+++ src/compat.jl
@@ -1,0 +1,6 @@
+if !isdefined(Base, :Returns)
+    struct Returns{T}
+        value::T
+    end
+    (f::Returns)(args...; kwds...) = f.value
+end